Fix ELN build
Upstream tarballs no longer include pre-built manpages, pandoc is not
included in RHEL, and the md2->man process relies on pandoc features
that simpler converters (such as go-md2man) do not support.

If this will be shipped in RHEL 10, then the manpages could be provided
by building them on Fedora or RHEL+EPEL, tarballing them, and adding
that to the sources.  That could be done post-branching once the RHEL
version is locked in, but is not sustainable for ELN which tracks rawhide.